Fix removeItem functionality
The remove function on the elm side was slurping up the key
parameter without passing it through to the native remove
function, causing a runtime error. This change alters the remove
function to be a simple passthrough (like set) so that all params
are sent to the native remove function.